pyproject.toml
[tool.poetry]
name = "pass-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "pass_server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.0"
uvicorn = "^0.23.2"
pydantic = "^2.1.1"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
